Tests/LibWeb/Text/expected/css/namespace-rules.txt

Can insert namespace rule into sheet with only @namespace and @import rules: PASS
Cannot insert namespace rule into sheet with existing rules: PASS
